+++ src/bin/club-search.ts
@@ -10,10 +10,11 @@ const log = console.log;
 
 export const program = commander
     .description(
-        `Search through clubhouse stories. Arguments (non-flag/options) will
-  be passed to Clubhouse story search API as search operators. Note that passing search
-  operators and options (e.g. --owner yourself) will use the options as extra filtering
-  in the client.
+        `Search through clubhouse stories. Arguments (non-flag/options) will be
+  passed to Clubhouse story search API as search operators. Passing '%self%' as
+  a search operator argument will be replaced ny your mention name. Note that
+  passing search operators and options (e.g. --owner yourself) will use the
+  options as extra filtering in the client.
 
   Refer to https://help.clubhouse.io/hc/en-us/articles/360000046646-Search-Operators
   for more details about search operators.`